+++ pyrasterframes/src/main/python/tests/RasterFunctionsTests.py
@@ -24,7 +24,10 @@ from pyrasterframes.rf_types import Tile
 from pyspark import Row
 from pyspark.sql.functions import *
 
+import numpy as np
+from numpy.testing import assert_equal
 
+from unittest import skip
 from . import TestEnvironment
 
 
@@ -103,7 +106,6 @@ class RasterFunctions(TestEnvironment):
     def test_agg_local_mean(self):
         from pyspark.sql import Row
         from pyrasterframes.rf_types import Tile
-        import numpy as np
 
         # this is really testing the nodata propagation in the agg  local summation
         ct = CellType.int8().with_no_data_value(4)
@@ -221,38 +223,131 @@ class RasterFunctions(TestEnvironment):
                                      rf_local_greater_int(self.rf.tile, 25000),
                                      "uint8"),
                                  lit(mask_value)).alias('mask'))
-        rf2 = rf1.select(rf1.tile, rf_mask_by_value(rf1.tile, rf1.mask, lit(mask_value)).alias('masked'))
+        rf2 = rf1.select(rf1.tile, rf_mask_by_value(rf1.tile, rf1.mask, lit(mask_value), False).alias('masked'))
         result = rf2.agg(rf_agg_no_data_cells(rf2.tile) < rf_agg_no_data_cells(rf2.masked)) \
             .collect()[0][0]
         self.assertTrue(result)
 
-        rf3 = rf1.select(rf1.tile, rf_inverse_mask_by_value(rf1.tile, rf1.mask, lit(mask_value)).alias('masked'))
-        result = rf3.agg(rf_agg_no_data_cells(rf3.tile) < rf_agg_no_data_cells(rf3.masked)) \
-            .collect()[0][0]
-        self.assertTrue(result)
+        # note supplying a `int` here, not a column to mask value
+        rf3 = rf1.select(
+            rf1.tile,
+            rf_inverse_mask_by_value(rf1.tile, rf1.mask, mask_value).alias('masked'),
+            rf_mask_by_value(rf1.tile, rf1.mask, mask_value, True).alias('masked2'),
+        )
+        result = rf3.agg(
+            rf_agg_no_data_cells(rf3.tile) < rf_agg_no_data_cells(rf3.masked),
+            rf_agg_no_data_cells(rf3.tile) < rf_agg_no_data_cells(rf3.masked2),
+        ) \
+            .first()
+        self.assertTrue(result[0])
+        self.assertTrue(result[1])  # inverse mask arg gives equivalent result
+
+        result_equiv_tiles = rf3.select(rf_for_all(rf_local_equal(rf3.masked, rf3.masked2))).first()[0]
+        self.assertTrue(result_equiv_tiles)  # inverse fn and inverse arg produce same Tile
+
+    def test_mask_by_values(self):
+
+        tile = Tile(np.random.randint(1, 100, (5, 5)), CellType.uint8())
+        mask_tile = Tile(np.array(range(1, 26), 'uint8').reshape(5, 5))
+        expected_diag_nd = Tile(np.ma.masked_array(tile.cells, mask=np.eye(5)))
+
+        df = self.spark.createDataFrame([Row(t=tile, m=mask_tile)]) \
+            .select(rf_mask_by_values('t', 'm', [0, 6, 12, 18, 24]))  # values on the diagonal
+        result0 = df.first()
+        # assert_equal(result0[0].cells, expected_diag_nd)
+        self.assertTrue(result0[0] == expected_diag_nd)
+
+    def test_mask_bits(self):
+        t = Tile(42 * np.ones((4, 4), 'uint16'), CellType.uint16())
+        # with a varitey of known values
+        mask = Tile(np.array([
+            [1, 1, 2720, 2720],
+            [1, 6816, 6816, 2756],
+            [2720, 2720, 6900, 2720],
+            [2720, 6900, 6816, 1]
+        ]), CellType('uint16raw'))
+
+        df = self.spark.createDataFrame([Row(t=t, mask=mask)])
+
+        # removes fill value 1
+        mask_fill_df = df.select(rf_mask_by_bit('t', 'mask', 0, True).alias('mbb'))
+        mask_fill_tile = mask_fill_df.first()['mbb']
+
+        self.assertTrue(mask_fill_tile.cell_type.has_no_data())
+
+        self.assertTrue(
+            mask_fill_df.select(rf_data_cells('mbb')).first()[0],
+            16 - 4
+        )
+
+        # mask out 6816, 6900
+        mask_med_hi_cir = df.withColumn('mask_cir_mh',
+                                        rf_mask_by_bits('t', 'mask', 11, 2, [2, 3])) \
+            .first()['mask_cir_mh'].cells
+
+        self.assertEqual(
+            mask_med_hi_cir.mask.sum(),
+            5
+        )
+
+    @skip('Issue #422 https://github.com/locationtech/rasterframes/issues/422')
+    def test_mask_and_deser(self):
+        # duplicates much of test_mask_bits but
+        t = Tile(42 * np.ones((4, 4), 'uint16'), CellType.uint16())
+        # with a varitey of known values
+        mask = Tile(np.array([
+            [1, 1, 2720, 2720],
+            [1, 6816, 6816, 2756],
+            [2720, 2720, 6900, 2720],
+            [2720, 6900, 6816, 1]
+        ]), CellType('uint16raw'))
+
+        df = self.spark.createDataFrame([Row(t=t, mask=mask)])
+
+        # removes fill value 1
+        mask_fill_df = df.select(rf_mask_by_bit('t', 'mask', 0, True).alias('mbb'))
+        mask_fill_tile = mask_fill_df.first()['mbb']
+
+        self.assertTrue(mask_fill_tile.cell_type.has_no_data())
+
+        # Unsure why this fails. mask_fill_tile.cells is all 42 unmasked.
+        self.assertEqual(mask_fill_tile.cells.mask.sum(), 4,
+                         f'Expected {16 - 4} data values but got the masked tile:'
+                         f'{mask_fill_tile}'
+                         )
 
     def test_mask(self):
         from pyspark.sql import Row
         from pyrasterframes.rf_types import Tile, CellType
-        import numpy as np
 
         np.random.seed(999)
-        ma = np.ma.array(np.random.randint(0, 10, (5, 5), dtype='int8'), mask=np.random.rand(5, 5) > 0.7)
+        # importantly exclude 0 from teh range because that's the nodata value for the `data_tile`'s cell type
+        ma = np.ma.array(np.random.randint(1, 10, (5, 5), dtype='int8'), mask=np.random.rand(5, 5) > 0.7)
         expected_data_values = ma.compressed().size
         expected_no_data_values = ma.size - expected_data_values
         self.assertTrue(expected_data_values > 0, "Make sure random seed is cooperative ")
         self.assertTrue(expected_no_data_values > 0, "Make sure random seed is cooperative ")
 
-        df = self.spark.createDataFrame([
-            Row(t=Tile(np.ones(ma.shape, ma.dtype)), m=Tile(ma))
-        ])
+        data_tile = Tile(np.ones(ma.shape, ma.dtype), CellType.uint8())
 
-        df = df.withColumn('masked_t', rf_mask('t', 'm'))
+        df = self.spark.createDataFrame([Row(t=data_tile, m=Tile(ma))]) \
+            .withColumn('masked_t', rf_mask('t', 'm'))
+
         result = df.select(rf_data_cells('masked_t')).first()[0]
-        self.assertEqual(result, expected_data_values)
+        self.assertEqual(result, expected_data_values,
+                         f"Masked tile should have {expected_data_values} data values but found: {df.select('masked_t').first()[0].cells}."
+                         f"Original data: {data_tile.cells}"
+                         f"Masked by {ma}")
 
         nd_result = df.select(rf_no_data_cells('masked_t')).first()[0]
         self.assertEqual(nd_result, expected_no_data_values)
+
+        # deser of tile is correct
+        self.assertEqual(
+            df.select('masked_t').first()[0].cells.compressed().size,
+            expected_data_values
+        )
+
 
     def test_resample(self):
         from pyspark.sql.functions import lit
@@ -323,13 +418,9 @@ class RasterFunctions(TestEnvironment):
         # Look for the PNG magic cookie
         self.assertEqual(png_bytes[0:8], bytearray([0x89, 0x50, 0x4E, 0x47, 0x0D, 0x0A, 0x1A, 0x0A]))
 
-
-
-
     def test_rf_interpret_cell_type_as(self):
         from pyspark.sql import Row
         from pyrasterframes.rf_types import Tile
-        import numpy as np
 
         df = self.spark.createDataFrame([
             Row(t=Tile(np.array([[1, 3, 4], [5, 0, 3]]), CellType.uint8().with_no_data_value(5)))
@@ -344,11 +435,12 @@ class RasterFunctions(TestEnvironment):
     def test_rf_local_data_and_no_data(self):
         from pyspark.sql import Row
         from pyrasterframes.rf_types import Tile
-        import numpy as np
-        from numpy.testing import assert_equal
 
-        t = Tile(np.array([[1, 3, 4], [5, 0, 3]]), CellType.uint8().with_no_data_value(5))
-        #note the convert is due to issue #188
+        nd = 5
+        t = Tile(
+            np.array([[1, 3, 4], [nd, 0, 3]]),
+            CellType.uint8().with_no_data_value(nd))
+        # note the convert is due to issue #188
         df = self.spark.createDataFrame([Row(t=t)])\
             .withColumn('lnd', rf_convert_cell_type(rf_local_no_data('t'), 'uint8')) \
             .withColumn('ld',  rf_convert_cell_type(rf_local_data('t'),    'uint8'))
@@ -359,3 +451,32 @@ class RasterFunctions(TestEnvironment):
 
         result_d = result['ld']
         assert_equal(result_d.cells, np.invert(t.cells.mask))
+
+    def test_rf_local_is_in(self):
+        from pyspark.sql.functions import lit, array, col
+        from pyspark.sql import Row
+
+        nd = 5
+        t = Tile(
+            np.array([[1, 3, 4], [nd, 0, 3]]),
+            CellType.uint8().with_no_data_value(nd))
+        # note the convert is due to issue #188
+        df = self.spark.createDataFrame([Row(t=t)]) \
+            .withColumn('a', array(lit(3), lit(4))) \
+            .withColumn('in2', rf_convert_cell_type(
+                rf_local_is_in(col('t'), array(lit(0), lit(4))),
+                'uint8')) \
+            .withColumn('in3', rf_convert_cell_type(rf_local_is_in('t', 'a'), 'uint8')) \
+            .withColumn('in4', rf_convert_cell_type(
+                rf_local_is_in('t', array(lit(0), lit(4), lit(3))),
+                'uint8')) \
+            .withColumn('in_list', rf_convert_cell_type(rf_local_is_in(col('t'), [4, 1]), 'uint8'))
+
+        result = df.first()
+        self.assertEqual(result['in2'].cells.sum(), 2)
+        assert_equal(result['in2'].cells, np.isin(t.cells, np.array([0, 4])))
+        self.assertEqual(result['in3'].cells.sum(), 3)
+        self.assertEqual(result['in4'].cells.sum(), 4)
+        self.assertEqual(result['in_list'].cells.sum(), 2,
+                         "Tile value {} should contain two 1s as: [[1, 0, 1],[0, 0, 0]]"
+                         .format(result['in_list'].cells))